moved reporting_units_and_users to commands and deleted scheduled_tasks
