+++ dds_web/commands.py
@@ -564,8 +564,12 @@ def quarterly_usage():
     """Get the monthly usage for the units"""
 
     flask.current_app.logger.debug("Task: Collecting usage information from database.")
+
+    # Imports
+    # Installed
     import sqlalchemy
 
+    # Own
     from dds_web import db
     from dds_web.database import models
     from dds_web.utils import (
@@ -639,3 +643,87 @@ def quarterly_usage():
         flask.current_app.logger.exception(err)
         db.session.rollback()
         raise
+
+
+@click.command("reporting-units-and-users")
+@flask.cli.with_appcontext
+def reporting_units_and_users():
+    """At the start of every month, get number of units and users."""
+    # Imports
+    # # Installed
+    import csv
+    import flask_mail
+    import flask_sqlalchemy
+    import pathlib
+
+    # Own
+    from dds_web import errors, utils
+    from dds_web.database.models import User, Unit
+
+    # Get current date
+    current_date: str = utils.timestamp(ts_format="%Y-%m-%d")
+
+    # Location of reporting file
+    reporting_file: pathlib.Path = pathlib.Path("/code/doc/reporting/dds-reporting.csv")
+
+    # Error default
+    error: str = None
+
+    # Get email address
+    recipient: str = flask.current_app.config.get("MAIL_DDS")
+    default_subject: str = "DDS Unit / User report"
+    default_body: str = f"This email contains the DDS unit- and user statistics. The data was collected on: {current_date}."
+    error_subject: str = f"Error in {default_subject}"
+    error_body: str = "The cronjob 'reporting' experienced issues"
+
+    # Get units and count them
+    units: flask_sqlalchemy.BaseQuery = Unit.query
+    num_units: int = units.count()
+
+    # Count users
+    users: flask_sqlalchemy.BaseQuery = User.query
+    num_users_total: int = users.count()  # All users
+    num_superadmins: int = users.filter_by(type="superadmin").count()  # Super Admins
+    num_unit_users: int = users.filter_by(type="unituser").count()  # Unit Admins / Personnel
+    num_researchers: int = users.filter_by(type="researchuser").count()  # Researchers
+    num_users_excl_superadmins: int = num_users_total - num_superadmins
+
+    # Verify that sum is correct
+    if sum([num_superadmins, num_unit_users, num_researchers]) != num_users_total:
+        error: str = "Sum of number of users incorrect."
+    # Define csv file and verify that it exists
+    elif not reporting_file.exists():
+        error: str = "Could not find the csv file."
+
+    if error:
+        # Send email about error
+        file_error_msg: flask_mail.Message = flask_mail.Message(
+            subject=error_subject,
+            recipients=[recipient],
+            body=f"{error_body}: {error}",
+        )
+        utils.send_email_with_retry(msg=file_error_msg)
+        raise Exception(error)
+
+    # Add row with new info
+    with reporting_file.open(mode="a") as repfile:
+        writer = csv.writer(repfile)
+        writer.writerow(
+            [
+                current_date,
+                num_units,
+                num_researchers,
+                num_unit_users,
+                num_users_excl_superadmins,
+            ]
+        )
+
+    # Create email
+    msg: flask_mail.Message = flask_mail.Message(
+        subject=default_subject,
+        recipients=[recipient],
+        body=default_body,
+    )
+    with reporting_file.open(mode="r") as file:  # Attach file
+        msg.attach(filename=reporting_file.name, content_type="text/csv", data=file.read())
+    utils.send_email_with_retry(msg=msg)  # Send
